Fix Record class
* feat: Record class implementation

This commit stablished the record class implementation, responsible for storing all the data from the files created in the app.
This implementation is not final and is subject to changes in the future

* fix: Enums not conforming to PersistentModel (#3)

All Enums were chaged to extend from Codable so they conform to PersistentModel for SwiftData

* feat: Color inclusion in project (#4)

This commit makes the inclusion of the colors utilized through the project

* Feat: camera implementation (#5)

* feat: basic camera functionality implemented

* style: updated info.plist authorization request messages

* feat: Integrate PhotosPicker for gallery image selection

Implemented SwiftUI's PhotosPicker in CameraModel, allowing users to select images from their photo library to be used. Updated state management for selected items and images, and removed the unused createPhotoLibraryButton. Added and clarified comments in CapturedScreen for better code readability.

* fix: added missing file comments

* fix: fixed comments on camera-related files

* Create ContentView.swift

Recreated ContentView.swift that was overwritten by mistake.

* feat: Record class implementation

This commit stablished the record class implementation, responsible for storing all the data from the files created in the app.
This implementation is not final and is subject to changes in the future

---------

Co-authored-by: Pedro Wiezel <131211509+pwiez@users.noreply.github.com>